@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
 
-module.exports = (req, res, next) => {
+export default (req, res, next) => {
   try {
     console.log(req.headers);
     const token = req.headers.authorization.split(' ')[1];
